backend/requirements.txt
fastapi
requests
google-genai
httpx
python-dotenv
pytest